requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
python-multipart>=0.0.6
tenacity>=8.2.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4